chefWeb/requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
country_list==1.1.0
Django==5.0
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
facebook-sdk==3.1.0
Faker==28.0.0
gcloud==0.17.0
google-auth==2.34.0
googleapis-common-protos==1.65.0
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
httplib2==0.22.0
idna==3.8
itypes==1.2.0
Jinja2==3.1.4
jws==0.1.3
Markdown==3.7
MarkupSafe==2.1.5
oauth2client==3.0.0
Pillow==9.5.0
promise==2.3
protobuf==5.28.0rc3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.4.3
PyJWT==2.9.0
pyparsing==3.1.4
Pyrebase==3.0.27
python-dateutil==2.9.0.post0
python-jwt==2.0.1
pytz==2024.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
